refactor: rename REST control-plane routes for naming consistency
Apply the `<module>/<resource>` + spelled-out-verb conventions across /v1:
nest cms-drafts under cms/, rename curation/jobs to curator/jobs, spell
out ack -> acknowledge, drop the lone admin/ tier from audit-logs, and
unify feedback's reject -> dismiss (REST route + MCP tool feedback_dismiss
+ internal method names). Merge the two controllers that both mounted
v1/usage into one UsageController (routes unchanged). Updated agent-runtime
and dashboard consumers + integration tests; regenerated openapi.json and
docs-fixtures.

Breaking for REST clients pinned to old paths and MCP clients pinned to
feedback_reject.
