+++ packages/backend-core/docs-fixtures/mcp-tools.json
@@ -5322,7 +5322,7 @@
   {
     "name": "feedback_create",
     "title": "Feedback: Create",
-    "description": "Submit feedback about Munin. Stays local until an org admin approves it; rejection deletes the item. Set includeOrgName / includeUserName to attach attribution; both default false.",
+    "description": "Submit feedback about Munin. Stays local until an org admin approves it; dismissal deletes the item. Set includeOrgName / includeUserName to attach attribution; both default false.",
     "audiences": [
       "admin"
     ],
@@ -5433,9 +5433,9 @@
     }
   },
   {
-    "name": "feedback_reject",
-    "title": "Feedback: Reject",
-    "description": "Reject (delete) a pending feedback item. Nothing is sent to Munin.",
+    "name": "feedback_dismiss",
+    "title": "Feedback: Dismiss",
+    "description": "Dismiss (delete) a pending feedback item. Nothing is sent to Munin.",
     "audiences": [
       "admin"
     ],
