+++ packages/backend-core/src/control/audit-log.controller.ts
@@ -30,7 +30,7 @@ type ClientKind = 'sdk' | 'cli' | 'mcp' | 'unknown';
 const PAGE_SIZE_DEFAULT = 50;
 const PAGE_SIZE_MAX = 200;
 
-@Controller('v1/admin/audit-logs')
+@Controller('v1/audit-logs')
 @UseGuards(AuthGuard, ControlPlaneGuard, RoleGuard)
 @UseInterceptors(TenancyInterceptor, AuditInterceptor)
 @RequireRole('owner', 'admin')
