+++ packages/backend-core/src/modules/curator/curator-jobs.integration.test.ts
@@ -112,7 +112,7 @@ const skipReason = TEST_URL
   }
 
   it('enqueues a job, claims it, and acks it', async () => {
-    const enq = await call('/v1/curation/jobs', {
+    const enq = await call('/v1/curator/jobs', {
       method: 'POST',
       body: {
         jobUri: 'skill://kb/review-content',
@@ -127,7 +127,7 @@ const skipReason = TEST_URL
     expect(enqBody.alreadyPending).toBe(false);
     expect(enqBody.job.status).toBe('pending');
 
-    const claim = await call('/v1/curation/jobs/claim', {
+    const claim = await call('/v1/curator/jobs/claim', {
       method: 'POST',
       body: { holder: 'sidecar-test', limit: 5, leaseSeconds: 60 },
     });
@@ -138,7 +138,7 @@ const skipReason = TEST_URL
     expect(claimBody.items[0]?.attempts).toBe(1);
     expect(claimBody.items[0]?.leaseHolder).toBe('sidecar-test');
 
-    const ack = await call(`/v1/curation/jobs/${enqBody.job.id}/ack`, {
+    const ack = await call(`/v1/curator/jobs/${enqBody.job.id}/acknowledge`, {
       method: 'POST',
       body: { replyText: 'done', toolCalls: 3, totalTokens: 200 },
     });
@@ -148,7 +148,7 @@ const skipReason = TEST_URL
     expect(ackBody.lastReplyText).toBe('done');
     expect(ackBody.lastToolCalls).toBe(3);
 
-    const reclaim = await call('/v1/curation/jobs/claim', {
+    const reclaim = await call('/v1/curator/jobs/claim', {
       method: 'POST',
       body: { holder: 'sidecar-test' },
     });
@@ -157,7 +157,7 @@ const skipReason = TEST_URL
   });
 
   it('returns existing job when same dedupeKey is enqueued twice', async () => {
-    const first = await call('/v1/curation/jobs', {
+    const first = await call('/v1/curator/jobs', {
       method: 'POST',
       body: {
         jobUri: 'skill://kb/review-content',
@@ -168,7 +168,7 @@ const skipReason = TEST_URL
     const firstBody = first.body as { job: { id: string }; alreadyPending: boolean };
     expect(firstBody.alreadyPending).toBe(false);
 
-    const second = await call('/v1/curation/jobs', {
+    const second = await call('/v1/curator/jobs', {
       method: 'POST',
       body: {
         jobUri: 'skill://kb/review-content',
@@ -183,18 +183,18 @@ const skipReason = TEST_URL
   });
 
   it('fail with retryable=true bumps next_attempt_at and stays pending', async () => {
-    const enq = await call('/v1/curation/jobs', {
+    const enq = await call('/v1/curator/jobs', {
       method: 'POST',
       body: { jobUri: 'skill://kb/review-content', userPrompt: 'will fail' },
     });
     const job = (enq.body as { job: { id: string } }).job;
 
-    await call('/v1/curation/jobs/claim', {
+    await call('/v1/curator/jobs/claim', {
       method: 'POST',
       body: { holder: 'sidecar-test' },
     });
 
-    const fail = await call(`/v1/curation/jobs/${job.id}/fail`, {
+    const fail = await call(`/v1/curator/jobs/${job.id}/fail`, {
       method: 'POST',
       body: { error: 'transient network error', retryable: true },
     });
@@ -206,24 +206,24 @@ const skipReason = TEST_URL
   });
 
   it('fail with retryable=false marks the job failed (not retried)', async () => {
-    const enq = await call('/v1/curation/jobs', {
+    const enq = await call('/v1/curator/jobs', {
       method: 'POST',
       body: { jobUri: 'skill://kb/review-content', userPrompt: 'permanent fail' },
     });
     const job = (enq.body as { job: { id: string } }).job;
 
-    await call('/v1/curation/jobs/claim', {
+    await call('/v1/curator/jobs/claim', {
       method: 'POST',
       body: { holder: 'sidecar-test' },
     });
 
-    const fail = await call(`/v1/curation/jobs/${job.id}/fail`, {
+    const fail = await call(`/v1/curator/jobs/${job.id}/fail`, {
       method: 'POST',
       body: { error: 'skill missing', retryable: false },
     });
     expect((fail.body as { status: string }).status).toBe('failed');
 
-    const reclaim = await call('/v1/curation/jobs/claim', {
+    const reclaim = await call('/v1/curator/jobs/claim', {
       method: 'POST',
       body: { holder: 'sidecar-test' },
     });
@@ -231,7 +231,7 @@ const skipReason = TEST_URL
   });
 
   it('retryable fail beyond maxAttempts marks the job dead', async () => {
-    const enq = await call('/v1/curation/jobs', {
+    const enq = await call('/v1/curator/jobs', {
       method: 'POST',
       body: { jobUri: 'skill://kb/review-content', userPrompt: 'eventual dead', maxAttempts: 2 },
     });
@@ -242,7 +242,7 @@ const skipReason = TEST_URL
       .set({ attempts: 2, leaseExpiresAt: null, leaseHolder: null })
       .where(sql`id = ${job.id}`);
 
-    const fail = await call(`/v1/curation/jobs/${job.id}/fail`, {
+    const fail = await call(`/v1/curator/jobs/${job.id}/fail`, {
       method: 'POST',
       body: { error: 'still broken', retryable: true },
     });
@@ -250,15 +250,15 @@ const skipReason = TEST_URL
   });
 
   it('isolates jobs across orgs (RLS)', async () => {
-    await call('/v1/curation/jobs', {
+    await call('/v1/curator/jobs', {
       method: 'POST',
       body: { jobUri: 'skill://kb/review-content', userPrompt: 'org A job' },
     });
 
-    const otherList = await call('/v1/curation/jobs', { key: otherAdminKey });
+    const otherList = await call('/v1/curator/jobs', { key: otherAdminKey });
     expect((otherList.body as { items: unknown[] }).items).toHaveLength(0);
 
-    const otherClaim = await call('/v1/curation/jobs/claim', {
+    const otherClaim = await call('/v1/curator/jobs/claim', {
       method: 'POST',
       key: otherAdminKey,
       body: { holder: 'sidecar-other' },
@@ -267,7 +267,7 @@ const skipReason = TEST_URL
   });
 
   it('rejects unknown skill:// URIs', async () => {
-    const res = await call('/v1/curation/jobs', {
+    const res = await call('/v1/curator/jobs', {
       method: 'POST',
       body: { jobUri: 'skill://kb/not-a-real-skill', userPrompt: 'x' },
     });
@@ -275,7 +275,7 @@ const skipReason = TEST_URL
   });
 
   it('rejects unknown task:// URIs', async () => {
-    const res = await call('/v1/curation/jobs', {
+    const res = await call('/v1/curator/jobs', {
       method: 'POST',
       body: { jobUri: 'task://web/run-arbitrary-code', userPrompt: 'x' },
     });
@@ -283,7 +283,7 @@ const skipReason = TEST_URL
   });
 
   it('rejects web-scrape jobs targeting the cloud metadata IP', async () => {
-    const res = await call('/v1/curation/jobs', {
+    const res = await call('/v1/curator/jobs', {
       method: 'POST',
       body: { jobUri: 'task://web/scrape-website', userPrompt: 'http://169.254.169.254/' },
     });
@@ -292,7 +292,7 @@ const skipReason = TEST_URL
   });
 
   it('rejects web-scrape jobs targeting localhost', async () => {
-    const res = await call('/v1/curation/jobs', {
+    const res = await call('/v1/curator/jobs', {
       method: 'POST',
       body: { jobUri: 'task://web/scrape-website', userPrompt: 'http://localhost:6379' },
     });
@@ -300,7 +300,7 @@ const skipReason = TEST_URL
   });
 
   it('rejects web-scrape jobs with an unsupported scheme', async () => {
-    const res = await call('/v1/curation/jobs', {
+    const res = await call('/v1/curator/jobs', {
       method: 'POST',
       body: { jobUri: 'task://web/scrape-website', userPrompt: 'ftp://example.com' },
     });
