+++ packages/dashboard-pages/src/pages/agent-setup-wizard.tsx
@@ -226,7 +226,7 @@ function WebsiteImportStatus({ jobId }: { jobId: string }) {
       if (cancelled) return;
       attempts += 1;
       try {
-        const res = await api<CuratorJobDto>(`/v1/curation/jobs/${jobId}`);
+        const res = await api<CuratorJobDto>(`/v1/curator/jobs/${jobId}`);
         if (cancelled) return;
         setJob(res);
         if (res.status !== 'pending' || attempts >= MAX_ATTEMPTS) {
